Implement WikiText byte-prediction evolution runner in Rust
- Added `wiki_runner.rs` to handle command-line interface and experiment execution.
- Introduced `WikiEnvironment` and `WikiAutomaton` structures in `wiki.rs` for managing WikiText data and automaton behavior.
- Implemented loading of WikiText datasets from Parquet files and handling of checkpoints.
- Created `run_pool` and `run_resumed_population` functions for managing population evolution and resuming from checkpoints.
- Added validation for command-line arguments and ensured compatibility with existing Python runner defaults.
- Included tests for CLI defaults, dataset loading, and population evolution.

+++ src/automaton.rs
@@ -173,8 +173,8 @@ impl MazeAutomaton {
         let input_code =
             ((self.internal_state as u32) << self.env_bits) | (env_local & self.env_mask) as u32;
         let output_code = self.genetic_code.get(input_code);
-        self.internal_state = output_code & self.state_mask;
-        let action = output_code >> self.state_bits;
+        self.internal_state = (output_code & self.state_mask as u16) as u8;
+        let action = (output_code >> self.state_bits) as u8;
 
         self.energy -= 1;
         self.attempt_action(action & self.resp_mask, maze);
